delete more generated files
