--- src/crossbars/blocking_task1/config.sby
@@ -1,18 +1,0 @@
-[options]
-
-mode bmc
-depth 100
-
-[engines]
-
-smtbmc
-
-[script]
-
-read -formal blocking.v
-
-prep -top blocking
-
-[files]
-
-blocking.v
