--- src/crossbars/blocking_with_spi_task1/blocking_with_spi_task1.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<testsuites>
-<testsuite timestamp="2024-05-08T21:56:20" hostname="c2s2-dev.ece.cornell.edu" package="blocking_with_spi" id="0" name="task1" tests="1" errors="1" failures="0" time="0" skipped="0">
-<properties>
-<property name="os" value="Linux"/>
-<property name="expect" value="PASS"/>
-<property name="status" value="ERROR"/>
-</properties>
-<testcase classname="task1" name="task1" time="0">
-<error type="ERROR"/>
-</testcase>
-<system-out>SBY 21:56:20 [blocking_with_spi_task1] Removing directory '/home/eyl47/c2s2_2024/c2s2_ip/src/crossbars/blocking_with_spi_task1'.
-SBY 21:56:20 [blocking_with_spi_task1] Copy '/home/eyl47/c2s2_2024/c2s2_ip/src/crossbars/blocking_with_spi.sv' to '/home/eyl47/c2s2_2024/c2s2_ip/src/crossbars/blocking_with_spi_task1/src/blocking_with_spi.sv'.
-SBY 21:56:20 [blocking_with_spi_task1] engine_0: smtbmc
-SBY 21:56:20 [blocking_with_spi_task1] base: starting process &quot;cd blocking_with_spi_task1/src; yosys -ql ../model/design.log ../model/design.ys&quot;
-SBY 21:56:20 [blocking_with_spi_task1] base: blocking_with_spi.sv:12: ERROR: syntax error, unexpected '[', expecting ',' or '=' or ')'
-SBY 21:56:20 [blocking_with_spi_task1] base: finished (returncode=1)
-SBY 21:56:20 [blocking_with_spi_task1] base: task failed. ERROR.
-SBY 21:56:20 [blocking_with_spi_task1] summary: Elapsed clock time [H:MM:SS (secs)]: 0:00:00 (0)
-SBY 21:56:20 [blocking_with_spi_task1] summary: Elapsed process time [H:MM:SS (secs)]: 0:00:00 (0)
-SBY 21:56:20 [blocking_with_spi_task1] summary: engine_0 (smtbmc) did not return a status
-SBY 21:56:20 [blocking_with_spi_task1] summary: engine_0 did not produce any traces
-SBY 21:56:20 [blocking_with_spi_task1] DONE (ERROR, rc=16)
-</system-out>
-<system-err>
-</system-err>
-</testsuite>
-</testsuites>
