--- src/crossbars/blocking_with_spi_task1/config.sby
@@ -1,17 +1,0 @@
-[options]
-
-mode bmc
-depth 100
-
-[engines]
-
-smtbmc
-
-[script]
-
-read_verilog -sv blocking_with_spi.sv
-prep -top blocking_with_spi
-
-[files]
-
-blocking_with_spi.sv
