--- src/crossbars/blocking_with_spi_task1/model/design_prep.ys
@@ -1,16 +1,0 @@
-# running in blocking_with_spi_task1/model/
-read_ilang design.il
-scc -select; simplemap; select -clear
-memory_nordff
-async2sync
-chformal -assume -early
-opt_clean
-formalff -setundef -clk2ff -ff2anyinit -hierarchy
-chformal -live -fair -cover -remove
-opt_clean
-check
-setundef -undriven -anyseq
-opt -fast
-rename -witness
-opt_clean
-write_rtlil ../model/design_prep.il
