--- src/crossbars/blocking_with_spi_task1/model/design_smt2.ys
@@ -1,7 +1,0 @@
-# running in blocking_with_spi_task1/model/
-read_ilang design_prep.il
-hierarchy -smtcheck
-formalff -assume
-dffunmap
-stat
-write_smt2 -wires design_smt2.smt2
