+++ src/crossbars/sim.py
@@ -17,3 +17,28 @@ def create_crossbar(nbits: int, nin: int, nout: int) -> tuple[callable, callable
         out_sel = _out_sel
 
     return crossbar, configure
+
+
+# Creates a crossbar with override bits
+def create_crossbar_with_spi(
+    nbits: int, nin: int, nout: int
+) -> tuple[callable, callable]:
+    in_sel = 0
+    out_sel = 0
+
+    def crossbar(inp: list[any]):
+        outputs = [0 for _ in range(nout)]
+        outputs[out_sel] = inp[in_sel]
+        # All outputs are zero except for the selected output
+        return outputs
+
+    def configure(_in_sel: int, _out_sel: int, input_spi: int, output_spi: int):
+        nonlocal in_sel, out_sel
+        in_sel = _in_sel
+        out_sel = _out_sel
+        if input_spi:
+            in_sel = 0
+        if output_spi:
+            out_sel = 0
+
+    return crossbar, configure
